resource_server/requirements.txt
Flask==1.1.1
PyJWT==1.7.1
cryptography==2.8
requests==2.23.0
PyYAML==5.3.1
pycryptodome==3.9.7
jwcrypto==0.7
oidcpy==0.0.3